requirements.txt
peewee
fastapi
uvicorn
pyjwt
bcrypt